requirements.txt
langgraph>=0.2.34
langchain>=0.3.0
openai>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.1
tenacity>=8.3.0
tiktoken>=0.7.0
rich>=13.7.0
typer>=0.12.3


